i added a prototype for the dicision making

+++ Game_Settings.py
@@ -1,4 +1,4 @@
 blind:float = 30
-startingMoney:float = 1000
+startingMoney:int = 1000
 minNbPlayer:int = 3
 maxNbPlayer:int = 6
